moved testStatus util
