+++ src/applications/static-pages/tests/application-status.cypress.spec.js
@@ -1,26 +1,4 @@
-import Timeouts from 'platform/testing/e2e/timeouts';
-import mockUser from './fixtures/mocks/mockUser';
-
-Cypress.Commands.add('testStatus', (page, url) => {
-  cy.visit(page);
-  cy.get('.sip-application-status').should('be.visible', {
-    timeout: Timeouts.slow,
-  });
-  cy.injectAxeThenAxeCheck();
-
-  cy.get('main a.usa-button-primary').should('have.attr', 'href', url);
-
-  cy.get('.usa-button-secondary')
-    .should('exist')
-    .then(button => {
-      cy.wrap(button).click();
-    });
-  cy.get('#start-over-modal-title').should(
-    'contain',
-    'Starting over will delete your in-progress application.',
-  );
-  cy.axeCheck();
-});
+import mockUser from 'platform/testing/e2e/mock-user';
 
 describe('Application Status Test', () => {
   it('Achieves the correct result per URL', () => {
