+++ src/platform/testing/e2e/mock-user.js
